Register based on DiscoveryInfo
tasks resources/ports range shouldn't be used by any discovery mechanism since
it is absolutely not ordered based on Mesos resource allocation order. In fact,
it is sorted based on port number.

This can have bad consequences when the app ask for more than one port AND the
agent reach the upper limit of the Mesos port range after allocating the first
port. The second allocated port is consequently lower than the first and we end
in a mess between Marathon and Consul.

By the way, the method used to populate based on ports names used DiscoveryInfo
so there was an additional mess because, again, it is not consitent with
resources/ports range.

Here we explicitely use DicoveryInfo as the source of truth regarding port
allocation, add the first port as the main service and iterate over the whole
array to populate services based on port names.

Signed-off-by: Pierre Cheynier <p.cheynier@criteo.com>